Rename native runner to builtin runner
